drm/tegra: dc: rgb: Move PCLK shifter programming to CRTC
Asus TF700T tablet uses TC358768 DPI to DSI bridge for display panel,
it requires to have a stable PCLK output before RGB encoder is enabled
because bridge starts to use PCLK at that point and bridge is programmed
before encoder is enabled. Hence the PCLK clock shifter must be programmed
when CRTC is enabled, otherwise output clock is not usable and bridge
hangs because of it. Move the shifter programming out of RGB encoder
into CRTC.

Tested-by: Maxim Schwalm <maxim.schwalm@gmail.com>
Signed-off-by: Dmitry Osipenko <digetx@gmail.com>