usb: phy: tegra: Correct definition of B_SESS_VLD_WAKEUP_EN
The B_SESS_VLD_WAKEUP_EN bit 6 was added by a mistake in a previous
commit, it corresponds to B_SESS_END_WAKEUP_EN. The B_VBUS_VLD_WAKEUP_EN
bit doesn't exist at all and B_SESS_VLD_WAKEUP_EN needs to be in place of
it. We don't utilize B-sensors, so it never was a problem, nevertheless
correct the definition of the bits.

Fixes: 35192007d28d ("usb: phy: tegra: Support waking up from a low power mode")
Signed-off-by: Dmitry Osipenko <digetx@gmail.com>