drm/tegra: Temporary pixel format fix for i927
Pixel format of the panel on this device is ABGR, not the typical RGBA
format. There's no way to switch the pixel format in panel as there's no
documenation, so do this in DRM driver instead.

The proper solution would require checking MEDIA_BUS_FMT_* values in DRM
driver and setting this value correctly in panel driver.

Signed-off-by: Sergey Larin <cerg2010cerg2010@mail.ru>