scftorture: Provide additional debug for memory-ordering diagnostics
The scftorture_invoke_one() function checks that stores carried out in
scf_handler() are visible when this function actually executed and when
scftorture_invoke_one() waited for the handler to complete.  However,
the diagnostic gives no clue as to which variant of smp_call_function()
was involved.  This commit therefore prints out this information when
the store was not visible to scftorture_invoke_one().

Signed-off-by: Paul E. McKenney <paulmck@kernel.org>