memory: tegra: Don't error out DMA flushing
DMA flushing may fail if hardware unit is power-gated. We need to reset
hardware unit and its memory clients in a certain order, but it becomes
complicated when we need to power-up hardware domain for the first time
since resets are requested in a special order. We assumed that power
domains are initially turned on all devices, but this could be a wrong
assumption.

Signed-off-by: Dmitry Osipenko <digetx@gmail.com>