Makefile: Enable -Wimplicit-fallthrough for Clang
With the recent fixes for fallthrough warnings, it is now possible to
enable -Wimplicit-fallthrough for Clang.

Notice that since we have adopted the new pseudo-keyword macro
fallthrough; we also want to avoid having more /* fall through */
comments being introduced. Contrary to GCC, Clang doesn't recognize
any comments as implicit fall-through markings when the
-Wimplicit-fallthrough option is enabled. So, in order to avoid
more comments being introduced, we need to use the option
-Wimplicit-fallthrough=5 for GCC.

Co-developed-by: Kees Cook <keescook@chromium.org>
Signed-off-by: Kees Cook <keescook@chromium.org>
Signed-off-by: Gustavo A. R. Silva <gustavoars@kernel.org>