mm/vmscan: put the redirtied MADV_FREE pages back to anonymous LRU list
Patch series "Cleanup and fixup for vmscan".

This series contains cleanups to remove unneeded return value, misleading
comment and so on.  Also this puts the redirtied MADV_FREE pages back to
anonymous LRU list.  More details can be found in the respective
changelogs.

This patch (of 5):

If the MADV_FREE pages are redirtied before they could be reclaimed, put
the pages back to anonymous LRU list by setting SwapBacked flag and the
pages will be reclaimed in normal swapout way.  Otherwise MADV_FREE pages
won't be reclaimed as expected.

Link: https://lkml.kernel.org/r/20210710100329.49174-1-linmiaohe@huawei.com
Link: https://lkml.kernel.org/r/20210710100329.49174-2-linmiaohe@huawei.com
Fixes: 802a3a92ad7a ("mm: reclaim MADV_FREE pages")
Signed-off-by: Miaohe Lin <linmiaohe@huawei.com>
Cc: Johannes Weiner <hannes@cmpxchg.org>
Cc: Vlastimil Babka <vbabka@suse.cz>
Cc: Michal Hocko <mhocko@suse.com>
Cc: Jens Axboe <axboe@kernel.dk>
Cc: Joonsoo Kim <iamjoonsoo.kim@lge.com>
Cc: Alex Shi <alexs@kernel.org>
Cc: Alistair Popple <apopple@nvidia.com>
Cc: Matthew Wilcox <willy@infradead.org>
Cc: Minchan Kim <minchan@kernel.org>
Cc: David Hildenbrand <david@redhat.com>
Cc: Shaohua Li <shli@fb.com>
Cc: Hillf Danton <hillf.zj@alibaba-inc.com>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>
Signed-off-by: Stephen Rothwell <sfr@canb.auug.org.au>