KVM: nSVM: use vmcb_save_area_cached in nested_vmcb_valid_sregs()
Now that struct vmcb_save_area_cached contains the required
vmcb fields values (done in nested_copy_vmcb_save_to_cache()),
check them to see if they are correct in nested_vmcb_valid_sregs().

Since we are always checking the data in the cache, it is enough
to keep only the vcpu as parameter.

Signed-off-by: Emanuele Giuseppe Esposito <eesposit@redhat.com>
Message-Id: <20211011143702.1786568-5-eesposit@redhat.com>
Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>