dm integrity: flush journal on suspend
There was a problem that the user burned dm-integrity image on CDROM
and could not activate it becaues it had non-empty journal.

Fixes this by flushing the journal on suspend.

Note that we deliberately didn't flush the journal on suspend, so that
the journal replay code would be tested. However, the dm-integrity
code is 5 years old now, so that journal replay is well-tested, and we
can make this change now.

Signed-off-by: Mikulas Patocka <mpatocka@redhat.com>
Signed-off-by: Mike Snitzer <snitzer@kernel.org>