printk: Skip console drivers on PREEMPT_RT.
printk might be invoked in a context with disabled interrupts and or
preemption and additionally disables interrupts before it invokes the
console drivers. This behavior is not desired on PREEMPT_RT:

- The console driver are using spinlock_t based locking which become
  sleeping locks on PREEMPT_RT and must not be acquired with disabled
  interrupts (or preemption).

- The locks within the console drivers must remain sleeping locks and
  they must not disable interrupts. Printing (and polling for its
  completion) at 115200 baud on an UART takes too long for PREEMPT_RT
  in general and so raises the latency of the IRQ-off time of the system
  beyond acceptable levels.

Skip printing to the console as temporary workaround until the printing
threads and atomic consoles have been introduced or another solution which
is compatible with the PREEMPT_RT approach.

With this change, the user will not see any kernel message printed to the
console but can retrieve the printk buffer from userland, for example,
via the dmesg command. This allows enable PREEMPT_RT as a whole without
disabling printk and losing all kernel output.

Disable console printing on PREEMPT_RT.

Signed-off-by: Sebastian Andrzej Siewior <bigeasy@linutronix.de>
Reviewed-by: Petr Mladek <pmladek@suse.com>
Reviewed-by: John Ogness <john.ogness@linutronix.de>
[pmladek@suse.com: Re-format the commit message. dmesg is just an example.]
Signed-off-by: Petr Mladek <pmladek@suse.com>
Link: https://lore.kernel.org/r/Yt6zwP9xSdUhsoQ9@linutronix.de