x86/mm: Fix sparse warnings in untagged_ptr()
Linear Address Masking patchset triggered a lot of sparse warnings.

The root cause is that casting pointer to '__typeof__(*(ptr)) *' will
strip all sparse tags. The type has to be defined based on the pointer
type, not based on what the pointer points to.

Fix cast in untagged_ptr() and avoid __typeof__() usage in
get/put_user().

Fixes: 5744534bdae4 ("x86/uaccess: Provide untagged_addr() and remove tags before address check")
Reported-by: kernel test robot <lkp@intel.com>
Signed-off-by: Kirill A. Shutemov <kirill.shutemov@linux.intel.com>
Signed-off-by: Dave Hansen <dave.hansen@linux.intel.com>
Link: https://lore.kernel.org/all/20221116004353.15052-3-kirill.shutemov%40linux.intel.com